Move term find-related code to find/
