+++ src/vs/workbench/contrib/terminal/browser/find/terminalFindWidget.ts
