+++ src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts
@@ -289,3 +289,5 @@ setupTerminalCommands();
 setupTerminalMenus();
 
 registerColors();
+
+import './find/terminal.find.contribution';
