+++ src/vs/workbench/contrib/terminal/common/terminalStrings.ts
@@ -14,6 +14,10 @@ export const terminalStrings = {
 	doNotShowAgain: localize('doNotShowAgain', 'Do Not Show Again'),
 	currentSessionCategory: localize('currentSessionCategory', 'current session'),
 	previousSessionCategory: localize('previousSessionCategory', 'previous session'),
+	actionCategory: {
+		value: localize('terminalCategory', "Terminal"),
+		original: 'Terminal'
+	},
 	focus: {
 		value: localize('workbench.action.terminal.focus', "Focus Terminal"),
 		original: 'Focus Terminal'
